part4: Add exercies 4.15.-4.23.

+++ part4/bloglist/controllers/blogs.js
@@ -1,44 +1,65 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const body = new Blog(request.body)
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const body = request.body
+
+  const user = request.user
 
   if (!body.title || !body.url){
-    return response.status(400).end()
+    return response.status(400).json({ error: 'title or url missing' })
   }
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+
+  if( blog.user.toString() === user._id.toString()){
+    await Blog.findByIdAndDelete(request.params.id)
+    return response.status(204).end()
+  }
+
+  response.status(404).json({ error: 'user invalid' })
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', userExtractor, async (request, response) => {
   const newBlogData = request.body
 
-  const updatedBlog = await Blog.findByIdAndUpdate(
-    request.params.id,
-    newBlogData,
-    { new: true }
-  )
+  const user = request.user
 
-  response.status(201).json(updatedBlog)
+  const blog = await Blog.findById(request.params.id)
+
+  if( blog.user.toString() === user._id.toString()){
+    const updatedBlog = await Blog.findByIdAndUpdate(
+      request.params.id,
+      newBlogData,
+      { new: true }
+    )
+    return response.status(201).json(updatedBlog)
+  }
+
+  response.status(404).json({ error: 'user invalid' })
 })
 
 module.exports = blogsRouter
